pyproject.toml
[project]
name = "website"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "wordpress-api>=1.2.9",
]
